Removed Spring, added KOIN.
