+++ build.gradle
@@ -1,11 +1,12 @@
 buildscript {
     ext.ktor_version = '1.0.1'
+    ext.koin_version = '1.0.2'
     repositories {
         mavenCentral()
+        jcenter()
     }
     dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.0.3.RELEASE")
-        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.0.1")
+        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:$ktor_version")
     }
 }
 
@@ -13,29 +14,27 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.3.10'
 }
 
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
-
 group 'ru.orbot90'
 version '1.0-SNAPSHOT'
 
 
 jar {
     manifest {
-        attributes 'Main-Class': 'ru.orbot90.com.coc.gmhelper.core.HelperApplication'
+        attributes 'Main-Class': 'ru.orbot90.com.coc.gmhelper.core.HelperApp'
     }
 }
 
 repositories {
     mavenCentral()
+    jcenter()
 }
 
 dependencies {
-    compile("org.springframework.boot:spring-boot-starter-web")
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8",
             "org.jetbrains.kotlin:kotlin-reflect",
             "io.ktor:ktor-server-netty:$ktor_version",
-            "io.ktor:ktor-jackson:$ktor_version"
+            "io.ktor:ktor-jackson:$ktor_version",
+            "org.koin:koin-ktor:$koin_version"
     testCompile("junit:junit")
 }
 
@@ -46,7 +45,7 @@ compileTestKotlin {
     kotlinOptions.jvmTarget = "1.8"
 }
 
-task buildDockerImage(dependsOn: bootJar, type: Exec) {
+task buildDockerImage(type: Exec) { //TODO: dependsOn: bootJar -> excluded, need to fix docker image
     workingDir "."
     commandLine "docker", "build", "-t", "orbot90/cocserver", "."
 }
