+++ src/main/kotlin/ru/orbot90/com/coc/gmhelper/core/controller/DiceRollController.kt
@@ -1,20 +1,18 @@
 package ru.orbot90.com.coc.gmhelper.core.controller
 
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.web.bind.annotation.*
-import ru.orbot90.com.coc.gmhelper.core.dice.DiceRoller
-import ru.orbot90.com.coc.gmhelper.core.model.RollRequest
+import io.ktor.application.call
+import io.ktor.request.receive
+import io.ktor.response.respond
+import io.ktor.routing.Route
+import io.ktor.routing.post
+import ru.orbot90.com.coc.gmhelper.core.model.TestRequest
+import ru.orbot90.com.coc.gmhelper.core.test.SkillTestPerformer
 
-@RestController
-@RequestMapping("/diceroll")
-class DiceRollController {
-
-    @Autowired
-    lateinit var diceRoller: DiceRoller
-
-    @PostMapping
-    @CrossOrigin(origins = arrayOf("http://localhost:4200"))
-    fun rollDice(@RequestBody request: RollRequest): IntArray {
-        return this.diceRoller.rollDice(request.diceCount, request.diceFacesCount)
+fun Route.performTestController(skillTestPerformer: SkillTestPerformer){
+    post("/performtest"){
+        val request = call.receive<TestRequest>()
+        val testResult = skillTestPerformer.performSkillTest(request.skillValue, request.bonusDice,
+                request.penaltyDice).name
+        call.respond("{\"testResult\": \"$testResult\"}")
     }
 }
